refactor: Section/Field, SummaryCard, CollapsibleText を汎用UIとして切り出す (#152, #153, #154)
- ConfigPage.tsx から Section/Field を components/ui/Section.tsx に分離
- MetricsPage.tsx から SummaryCard を components/ui/SummaryCard.tsx に分離
- SessionDetail.tsx から CollapsibleText を components/ui/CollapsibleText.tsx に分離
