+++ frontend/src/components/MetricsPage.tsx
@@ -14,6 +14,7 @@ import {
 import { api } from "../api/client";
 import type { MetricsSummary, MetricRow, MetricEvent } from "../api/client";
 import { useNavigate } from "react-router-dom";
+import { SummaryCard } from "./ui/SummaryCard";
 
 type TimeRange = "1h" | "6h" | "24h" | "7d" | "all";
 
@@ -940,11 +941,3 @@ function EventSummaryText({ event }: { event: MetricEvent }) {
   }
 }
 
-function SummaryCard({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="bg-white rounded-lg border border-gray-200 p-3">
-      <div className="text-xs text-gray-500">{label}</div>
-      <div className="text-lg font-bold text-gray-900 mt-0.5">{value}</div>
-    </div>
-  );
-}
