+++ frontend/src/components/SessionDetail.tsx
@@ -9,6 +9,7 @@ import {
   Code, Eye,
 } from "lucide-react";
 import { Modal } from "./ui/Modal";
+import { CollapsibleText } from "./ui/CollapsibleText";
 import type { Session } from "../types/session";
 import { api, type DiffFile } from "../api/client";
 import { StatusDot } from "./StatusBadge";
@@ -754,40 +755,6 @@ function SystemEventView({ item }: { item: Extract<StreamDisplayItem, { kind: "s
   );
 }
 
-const COLLAPSE_LINE_THRESHOLD = 20;
-
-function CollapsibleText({ text }: { text: string }) {
-  const [expanded, setExpanded] = useState(false);
-  const lineCount = text.split("\n").length;
-  const shouldCollapse = lineCount > COLLAPSE_LINE_THRESHOLD;
-
-  if (!shouldCollapse || expanded) {
-    return (
-      <div>
-        <div className="prose prose-xs max-w-none prose-pre:bg-gray-100 prose-pre:text-gray-800 prose-code:text-pink-600">
-          <Markdown remarkPlugins={[remarkGfm]}>{text}</Markdown>
-        </div>
-        {shouldCollapse && (
-          <button onClick={() => setExpanded(false)} className="text-xs text-blue-500 hover:text-blue-700 mt-1">
-            折りたたむ
-          </button>
-        )}
-      </div>
-    );
-  }
-
-  const preview = text.split("\n").slice(0, 5).join("\n");
-  return (
-    <div>
-      <div className="prose prose-xs max-w-none prose-pre:bg-gray-100 prose-pre:text-gray-800 prose-code:text-pink-600 relative overflow-hidden max-h-24">
-        <Markdown remarkPlugins={[remarkGfm]}>{preview}</Markdown>
-      </div>
-      <button onClick={() => setExpanded(true)} className="text-xs text-blue-500 hover:text-blue-700 mt-1">
-        続きを表示 ({lineCount} 行)
-      </button>
-    </div>
-  );
-}
 
 function StreamDisplayItemView({ item, onAnswer, sessionId, escalationId, escalationTimedOut, escalationTimeoutSeconds, onEscalationResponded }: {
   item: StreamDisplayItem;
