+++ frontend/src/components/ui/SummaryCard.tsx
@@ -1,0 +1,8 @@
+export function SummaryCard({ label, value }: { label: string; value: string }) {
+  return (
+    <div className="bg-white rounded-lg border border-gray-200 p-3">
+      <div className="text-xs text-gray-500">{label}</div>
+      <div className="text-lg font-bold text-gray-900 mt-0.5">{value}</div>
+    </div>
+  );
+}
